+++ src/Providers.tsx
@@ -12,7 +12,7 @@ import { ExtrinsicsProvider } from 'contexts/Extrinsics';
 import { FastUnstakeProvider } from 'contexts/FastUnstake';
 import { FiltersProvider } from 'contexts/Filters';
 import { LedgerHardwareProvider } from 'contexts/Hardware/Ledger/LedgerHardware';
-import { VaultAccountsProvider } from 'contexts/Hardware/Vault';
+import { VaultAccountsProvider } from 'contexts/Hardware/Vault/VaultAccounts';
 import { HelpProvider } from 'contexts/Help';
 import { IdentitiesProvider } from 'contexts/Identities';
 import { MenuProvider } from 'contexts/Menu';
